--- EnigmaMachine/Engine/src/Engine/LoadEnigmaMachineHandler.java
@@ -1,19 +1,0 @@
-/*package Engine;
-
-
-import java.beans.EventHandler;
-import java.util.EventObject;
-
-public class LoadEnigmaMachineHandler {
-    private EventHandler loadEnigmaMachineHandler;
-
-    public void addLoadMachineHandler (EventHandler handler) {
-        if (handler != null && !loadEnigmaMachineHandler.(handler)) {
-            loadEnigmaMachineHandler.add(handler);
-        }
-    }
-
-    public void removeLoadMachineHandler(EventHandler handler) {
-        loadEnigmaMachineHandler.remove(handler);
-    }
-}*/
