+++ EnigmaMachine/Engine/src/Engine/StatisticsAndHistory/OriginalStringFormat.java
@@ -1,14 +1,14 @@
-package Engine;
+package Engine.StatisticsAndHistory;
 
 import java.io.Serializable;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class EncryptedStringFormat extends StringFormat<Character> implements Serializable {
+public class OriginalStringFormat extends StringFormat<Character> implements Serializable {
 
-    public EncryptedStringFormat(List<Character> elements) {
 
-        super(elements, StringFormatType.ENCRYPTED_STRING);
+    public OriginalStringFormat(List<Character> originalString) {
+        super(originalString, StringFormatType.ORIGINAL_STRING);
     }
 
     @Override
