+++ EnigmaMachine/Engine/src/Engine/StatisticsAndHistory/StringFormat.java
@@ -1,4 +1,4 @@
-package Engine;
+package Engine.StatisticsAndHistory;
 
 import java.io.Serializable;
 import java.util.List;
