+++ EnigmaMachine/Engine/src/Engine/StatisticsAndHistory/StringFormatType.java
@@ -1,4 +1,4 @@
-package Engine;
+package Engine.StatisticsAndHistory;
 
 import java.io.Serializable;
