+++ EnigmaMachine/Engine/src/EnigmaMachine/Settings/Sector.java
@@ -1,4 +1,4 @@
-package Engine;
+package EnigmaMachine.Settings;
 
 import EnigmaMachine.EnigmaMachine;
 import EnigmaMachineException.PluginBoardSettingsException;
